ii
git-svn-id: svn://192.168.0.99/MyUtility@5748 5ada3c48-002d-0410-850d-99865cc6fd0f

+++ stdwin32/stdwin32.cpp
@@ -32,7 +32,7 @@ static const wchar_t* Nil=L"";
 string stdGetModuleFileNameA(HINSTANCE hInst)
 {
 	LPSTR p = NULL;
-	size_t size = 64;
+	DWORD size = 64;
 	for(;;)
 	{
 		p = (LPSTR)realloc(p, size*sizeof(char));
@@ -48,7 +48,7 @@ string stdGetModuleFileNameA(HINSTANCE hInst)
 wstring stdGetModuleFileNameW(HINSTANCE hInst)
 {
 	LPWSTR p = NULL;
-	size_t size = 64;
+	DWORD size = 64;
 	for(;;)
 	{
 		p = (LPWSTR)realloc(p, size*sizeof(wchar_t));
